% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMR.PRESSO.R
\name{HDMR.PRESSO}
\alias{HDMR.PRESSO}
\title{HDMR.PRESSO}
\usage{
HDMR.PRESSO(fdata1, Methodname, cv.lambd)
}
\arguments{
\item{fdata1}{Conditional GWAS summary statistics for the target dataset. Output from function \emph{TransCondi}.}

\item{Methodname}{One of the penalty "Lasso","ElasticNet","MCP" and "SCAD".}

\item{cv.lambd}{A range of parameter \eqn{\lambda}.}
}
\value{
A list object, \emph{removeIV} is the index of invalid IVs. \emph{fdata2} is the new datasets removing invalid IVs.
}
\description{
Removing invalid IVs for TL.HDMR.
}
