% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMR.R
\name{HDMR}
\alias{HDMR}
\title{HDMR}
\usage{
HDMR(fdata1, cv.lambd, penalname, alpha)
}
\arguments{
\item{fdata1}{Output from function \emph{TransCondi}.}

\item{cv.lambd}{A range of parameter \eqn{\lambda}.}

\item{alpha}{The tuning parameter of "Lasso" and "ElasticNet". If the penalty term is "MCP" or "SCAD", ignore this parameter.}

\item{Methodname}{One of the penalty "Lasso","ElasticNet","MCP" and "SCAD".}
}
\value{
A vector, the results of causal effect estimation of each exposure on the outcome.
}
\description{
High-dimensional MR method.
}
