% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMR.TSD.R
\name{HDMR.TSD}
\alias{HDMR.TSD}
\title{HDMR.TSD}
\usage{
HDMR.TSD(fdata1, fdata_NK1, cv.lambd, Methodname, r.A0, C0)
}
\arguments{
\item{fdata1}{Conditional GWAS summary statistics for the target dataset. Output from function \emph{TransCondi}.}

\item{fdata_NK1}{A list of conditional GWAS summary statistics for source datasets. Output from function \emph{TransCondi}.}

\item{cv.lambd}{A range of parameter \eqn{\lambda}.}

\item{Methodname}{One of the penalty "Lasso","ElasticNet","MCP" and "SCAD".}

\item{r.A0}{The number of folds for target dataset in the TSD algorithm.}

\item{C0}{The threshold in the TSD algorithm.}
}
\value{
A list object, \emph{Ind.loc} is the index of the informative source datasets.
}
\description{
Transferable Source Detection for TL.HDMR.
}
