% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.HDMR.R
\name{TL.HDMR}
\alias{TL.HDMR}
\title{TL.HDMR}
\usage{
TL.HDMR(betaGX_A0, betaGY_A0, betaGX_AK, betaGY_AK, cv.lambd, Methodname)
}
\arguments{
\item{betaGX_A0}{A covariance-corrected matrix of G-X conditional association coefficients (nSNP*nexposure) from the target population, corresponding to the \eqn{\Omega_{Y}^{-1/2}\hat{A}^{(0)}} in the Hou et al.}

\item{betaGY_A0}{A covariance-corrected vector of G-Y conditional association coefficients (nSNP*1) from the target population, corresponding to the \eqn{\Omega_{Y}^{-1/2}\hat{\Gamma}^{(0)}} in the Hou et al.}

\item{betaGX_AK}{A covariance-corrected matrix of G-X conditional association coefficients (nSNP*nexposure) from the source population, corresponding to the \eqn{\Omega_{Y}^{-1/2}\hat{A}^{(k)}} in the Hou et al.}

\item{betaGY_AK}{A covariance-corrected vector of G-Y conditional association coefficients (nSNP*1) from the source population, corresponding to the \eqn{\Omega_{Y}^{-1/2}\hat{\Gamma}^{(k)}} in the Hou et al.}

\item{cv.lambd}{A range of parameter \eqn{\lambda}.}

\item{Methodname}{One of the penalty "Lasso","ElasticNet","MCP" and "SCAD".}
}
\value{
A list object, the results of causal effect estimation of each exposure on the outcome as well as the value lambda.
}
\description{
Transfer Learning for High-Dimensional Mendelian Randomization (Lei Hou)
}
