% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCondi.R
\name{TransCondi}
\alias{TransCondi}
\title{TransCondi}
\usage{
TransCondi(fdata)
}
\arguments{
\item{fdata}{A list of G-X and G-Y observed (marginal) associations.
Bx_obs is the G-X association coefficient (\eqn{\hat{\beta}_{p}});
By_obs is the G-Y association coefficient (\eqn{\hat{\beta}_{Y}});
Sigy_condi is the covariance matrix of \eqn{\hat{\beta}_{Y}},  corresponding to the \eqn{\Sigma_{Y}=R\sigma_{Y}\sigma_{Y}^{T}} in the Hou et al.;
LD_mat is the LD matrix of IVs (\eqn{R});
LD_inv is the inverse of LD matrix (\eqn{R^{-1}}).}
}
\value{
A list object, including G-X and G-Y conditional genetic associations (\eqn{\hat{\Gamma}} and \eqn{\hat{A}}) and their covariance matrix.
}
\description{
Transform marginal genetic associations into conditional associations
}
